+++ database/migrations/2024_06_06_065014_pengeluaran.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('pengeluaran', function (Blueprint $table) {
             $table->id('id_pengeluaran');
-            $table->float('jumlah_pengeluaran');
+            $table->double('jumlah_pengeluaran');
             $table->string('jenis_pengeluaran');
             $table->date('tanggal_pengeluaran');
             $table->string('keterangan_pengeluaran');
